fix : seq의 타입을 int에서 long으로 변경

+++ src/main/java/someone/alcoholic/repository/member/TmpMemberRepository.java
@@ -5,6 +5,6 @@ import someone.alcoholic.domain.member.TmpMember;
 
 import java.util.Optional;
 
-public interface TmpMemberRepository extends JpaRepository<TmpMember, Integer>{
+public interface TmpMemberRepository extends JpaRepository<TmpMember, Long>{
     Optional<TmpMember> findById(String id);
 }
